# Improve ContinueWith perf with NotOn* options