--- a/src/libraries/System.Private.CoreLib/src/System/Threading/Tasks/TaskContinuation.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Threading/Tasks/TaskContinuation.cs
@@ -335,7 +335,7 @@ namespace System.Threading.Tasks
             }
             // Otherwise, the final state of this task does not match the desired
             // continuation activation criteria; cancel it to denote this.
-            else continuationTask.InternalCancel(false);
+            else continuationTask.InternalCancel();
         }
 
         internal override Delegate[]? GetDelegateContinuationsForDebugger()