@stephentoub 2020-05-01
I removed the commit that removed the scheduler field from the continuation task; turns out our synchronous Wait inlining logic relied on that.  I'll look at removing it again separately.